+++ src/dipole.jl
@@ -58,7 +58,7 @@ function calc_dipole_geo(time)
     return SA[sθ * cφ, sθ * sφ, cθ]
 end
 
-calc_dipole_gei(time) = rotation(GEO, GEI, time) * calc_dipole_geo(time)
+calc_dipole_gei(time) = rotation(GEO => GEI, time) * calc_dipole_geo(time)
 
 """
     dipole_tilt(time)
@@ -70,7 +70,7 @@ positive when the north magnetic pole is tilted toward the Sun.
 """
 function dipole_tilt(time)
     gst, ra, dec = csundir(time)
-    dipole_gei = rotation(GEO, GEI, gst) * calc_dipole_geo(time)
+    dipole_gei = rotation(GEO => GEI, gst) * calc_dipole_geo(time)
     sun_gei = calc_sun_gei(ra, dec)
     # sin(μ) = dipole · sun
     return asin(dot(dipole_gei, sun_gei))